requirements.txt
configparser==5.0.2
flatdict==4.0.1
PyYAML==5.4.1
pytest==6.2.5